pyproject.toml
[project]
name = "gui-agent-screenshot-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Josue Figueroa", email = "poor.rican@pm.me" }
]
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
